+++ examples/scripts/ecm_CMAES.py
@@ -89,7 +89,7 @@ parameter_set.export_parameters(
 pybop.plot_dataset(dataset)
 
 # Plot the timeseries output
-pybop.quick_plot(problem, parameter_values=x, title="Optimised Comparison")
+pybop.quick_plot(problem, problem_inputs=x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot_convergence(optim)