+++ pybop/models/base_model.py
@@ -7,8 +7,7 @@ import numpy as np
 import pybamm
 
 from pybop import Dataset, Experiment, Parameters, ParameterSet
-
-Inputs = Dict[str, float]
+from pybop.parameters.parameter import Inputs
 
 
 @dataclass
@@ -65,7 +64,7 @@ class BaseModel:
         else:  # a pybop parameter set
             self._parameter_set = pybamm.ParameterValues(parameter_set.params)
 
-        self.parameters = None
+        self.parameters = Parameters()
         self.dataset = None
         self.signal = None
         self.additional_variables = []
@@ -73,10 +72,6 @@ class BaseModel:
         self.standard_parameters = {}
         self.param_check_counter = 0
         self.allow_infeasible_solutions = True
-
-    @property
-    def n_parameters(self):
-        return len(self.parameters)
 
     def build(
         self,
@@ -104,8 +99,8 @@ class BaseModel:
             The initial state of charge to be used in simulations.
         """
         self.dataset = dataset
-        self.parameters = parameters
-        if self.parameters is not None:
+        if parameters is not None:
+            self.parameters = parameters
             self.classify_and_update_parameters(self.parameters)
 
         if init_soc is not None:
@@ -227,8 +222,8 @@ class BaseModel:
             The initial state of charge to be used in simulations.
         """
         self.dataset = dataset
+
         if parameters is not None:
-            self.parameters = parameters
             self.classify_and_update_parameters(parameters)
 
         if init_soc is not None:
@@ -251,7 +246,7 @@ class BaseModel:
         # Clear solver and setup model
         self._solver._model_set_up = {}
 
-    def classify_and_update_parameters(self, parameters: Union[Parameters, Dict]):
+    def classify_and_update_parameters(self, parameters: Parameters):
         """
         Update the parameter values according to their classification as either
         'rebuild_parameters' which require a model rebuild and
@@ -259,10 +254,16 @@ class BaseModel:
 
         Parameters
         ----------
-        parameters : pybop.ParameterSet
+        parameters : pybop.Parameters
 
         """
-        parameter_dictionary = parameters.as_dict()
+        if parameters is None:
+            self.parameters = Parameters()
+        else:
+            self.parameters = parameters
+
+        parameter_dictionary = self.parameters.as_dict()
+
         rebuild_parameters = {
             param: parameter_dictionary[param]
             for param in parameter_dictionary
@@ -283,6 +284,9 @@ class BaseModel:
             self._unprocessed_parameter_set = self._parameter_set
             self.geometry = self.pybamm_model.default_geometry
 
+        # Update the list of parameter names and number of parameters
+        self._n_parameters = len(self.parameters)
+
     def reinit(
         self, inputs: Inputs, t: float = 0.0, x: Optional[np.ndarray] = None
     ) -> TimeSeriesState:
@@ -292,8 +296,7 @@ class BaseModel:
         if self._built_model is None:
             raise ValueError("Model must be built before calling reinit")
 
-        if not isinstance(inputs, dict):
-            inputs = self.parameters.as_dict(inputs)
+        inputs = self.parameters.verify(inputs)
 
         self._solver.set_up(self._built_model, inputs=inputs)
 
@@ -340,9 +343,8 @@ class BaseModel:
 
         Parameters
         ----------
-        inputs : dict or array-like
-            The input parameters for the simulation. If array-like, it will be
-            converted to a dictionary using the model's fit keys.
+        inputs : Inputs
+            The input parameters for the simulation.
         t_eval : array-like
             An array of time points at which to evaluate the solution.
 
@@ -356,6 +358,8 @@ class BaseModel:
         ValueError
             If the model has not been built before simulation.
         """
+        inputs = self.parameters.verify(inputs)
+
         if self._built_model is None:
             raise ValueError("Model must be built before calling simulate")
         else:
@@ -363,9 +367,6 @@ class BaseModel:
                 sol = self.solver.solve(self.built_model, t_eval=t_eval)
 
             else:
-                if not isinstance(inputs, dict):
-                    inputs = self.parameters.as_dict(inputs)
-
                 if self.check_params(
                     inputs=inputs,
                     allow_infeasible_solutions=self.allow_infeasible_solutions,
@@ -393,9 +394,8 @@ class BaseModel:
 
         Parameters
         ----------
-        inputs : dict or array-like
-            The input parameters for the simulation. If array-like, it will be
-            converted to a dictionary using the model's fit keys.
+        inputs : Inputs
+            The input parameters for the simulation.
         t_eval : array-like
             An array of time points at which to evaluate the solution and its
             sensitivities.
@@ -410,6 +410,7 @@ class BaseModel:
         ValueError
             If the model has not been built before simulation.
         """
+        inputs = self.parameters.verify(inputs)
 
         if self._built_model is None:
             raise ValueError("Model must be built before calling simulate")
@@ -418,9 +419,6 @@ class BaseModel:
                 raise ValueError(
                     "Cannot use sensitivies for parameters which require a model rebuild"
                 )
-
-            if not isinstance(inputs, dict):
-                inputs = self.parameters.as_dict(inputs)
 
             if self.check_params(
                 inputs=inputs,
@@ -440,7 +438,7 @@ class BaseModel:
                         (
                             sol[self.signal[0]].data.shape[0],
                             self.n_outputs,
-                            self.n_parameters,
+                            self._n_parameters,
                         )
                     )
 
@@ -478,10 +476,9 @@ class BaseModel:
 
         Parameters
         ----------
-        inputs : dict or array-like, optional
-            Input parameters for the simulation. If the input is array-like, it is converted
-            to a dictionary using the model's fitting keys. Defaults to None, indicating
-            that the default parameters should be used.
+        inputs : Inputs, optional
+            Input parameters for the simulation. Defaults to None, indicating that the
+            default parameters should be used.
         t_eval : array-like, optional
             An array of time points at which to evaluate the solution. Defaults to None,
             which means the time points need to be specified within experiment or elsewhere.
@@ -507,13 +504,13 @@ class BaseModel:
             if PyBaMM models are not supported by the current simulation method.
 
         """
+        inputs = self.parameters.verify(inputs)
+
         if not self.pybamm_model._built:
             self.pybamm_model.build_model()
 
         parameter_set = parameter_set or self._unprocessed_parameter_set
         if inputs is not None:
-            if not isinstance(inputs, dict):
-                inputs = self.parameters.as_dict(inputs)
             parameter_set.update(inputs)
 
         if self.check_params(
@@ -552,7 +549,7 @@ class BaseModel:
 
         Parameters
         ----------
-        inputs : dict
+        inputs : Inputs
             The input parameters for the simulation.
         allow_infeasible_solutions : bool, optional
             If True, infeasible parameter values will be allowed in the optimisation (default: True).
@@ -563,17 +560,7 @@ class BaseModel:
             A boolean which signifies whether the parameters are compatible.
 
         """
-        if inputs is not None:
-            if not isinstance(inputs, dict):
-                if isinstance(inputs, list):
-                    for entry in inputs:
-                        if not isinstance(entry, (int, float)):
-                            raise ValueError(
-                                "Expecting inputs in the form of a dictionary, numeric list"
-                                + f" or None, but received a list with type: {type(inputs)}"
-                            )
-                else:
-                    inputs = self.parameters.as_dict(inputs)
+        inputs = self.parameters.verify(inputs)
 
         return self._check_params(
             inputs=inputs, allow_infeasible_solutions=allow_infeasible_solutions
@@ -588,7 +575,7 @@ class BaseModel:
 
         Parameters
         ----------
-        inputs : dict
+        inputs : Inputs
             The input parameters for the simulation.
         allow_infeasible_solutions : bool, optional
             If True, infeasible parameter values will be allowed in the optimisation (default: True).
@@ -649,7 +636,7 @@ class BaseModel:
         """
         raise NotImplementedError
 
-    def approximate_capacity(self, x):
+    def approximate_capacity(self, inputs: Inputs):
         """
         Calculate a new estimate for the nominal capacity based on the theoretical energy density
         and an average voltage.
@@ -658,8 +645,8 @@ class BaseModel:
 
         Parameters
         ----------
-        x : array-like
-            An array of values representing the model inputs.
+        inputs : Inputs
+            The parameters that are the inputs of the model.
 
         Raises
         ------