+++ tests/unit/test_cost.py
@@ -9,6 +9,11 @@ class TestCosts:
     Class for tests cost functions
     """
 
+    # Define an invalid likelihood class for MAP tests
+    class InvalidLikelihood:
+        def __init__(self, problem, sigma0):
+            pass
+
     @pytest.fixture
     def model(self):
         return pybop.lithium_ion.SPM()
@@ -114,14 +119,39 @@ class TestCosts:
             base_cost.evaluateS1([0.5])
 
     @pytest.mark.unit
+    def test_error_in_cost_calculation(self, problem):
+        class RaiseErrorCost(pybop.BaseCost):
+            def _evaluate(self, inputs, grad=None):
+                raise ValueError("Error test.")
+
+            def _evaluateS1(self, inputs):
+                raise ValueError("Error test.")
+
+        cost = RaiseErrorCost(problem)
+        with pytest.raises(ValueError, match="Error in cost calculation: Error test."):
+            cost([0.5])
+        with pytest.raises(ValueError, match="Error in cost calculation: Error test."):
+            cost.evaluateS1([0.5])
+
+    @pytest.mark.unit
     def test_MAP(self, problem):
         # Incorrect likelihood
-        with pytest.raises(ValueError):
+        with pytest.raises(
+            ValueError,
+            match="An error occurred when constructing the Likelihood class:",
+        ):
             pybop.MAP(problem, pybop.SumSquaredError)
 
         # Incorrect construction of likelihood
-        with pytest.raises(ValueError):
-            pybop.MAP(problem, pybop.GaussianLogLikelihoodKnownSigma, sigma="string")
+        with pytest.raises(
+            ValueError,
+            match="An error occurred when constructing the Likelihood class: could not convert string to float: 'string'",
+        ):
+            pybop.MAP(problem, pybop.GaussianLogLikelihoodKnownSigma, sigma0="string")
+
+        # Incorrect likelihood
+        with pytest.raises(ValueError, match="must be a subclass of BaseLikelihood"):
+            pybop.MAP(problem, self.InvalidLikelihood, sigma0=0.1)
 
     @pytest.mark.unit
     def test_costs(self, cost):
@@ -159,7 +189,9 @@ class TestCosts:
             assert type(de) == np.ndarray
 
             # Test exception for non-numeric inputs
-            with pytest.raises(ValueError):
+            with pytest.raises(
+                TypeError, match="Inputs must be a dictionary or numeric."
+            ):
                 cost.evaluateS1(["StringInputShouldNotWork"])
 
             with pytest.warns(UserWarning) as record:
@@ -176,7 +208,7 @@ class TestCosts:
             assert cost.evaluateS1([0.01]) == (np.inf, cost._de)
 
         # Test exception for non-numeric inputs
-        with pytest.raises(ValueError):
+        with pytest.raises(TypeError, match="Inputs must be a dictionary or numeric."):
             cost(["StringInputShouldNotWork"])
 
         # Test treatment of simulations that terminated early
@@ -225,7 +257,9 @@ class TestCosts:
             assert cost([1.1]) == -np.inf
 
             # Test exception for non-numeric inputs
-            with pytest.raises(ValueError):
+            with pytest.raises(
+                TypeError, match="Inputs must be a dictionary or numeric."
+            ):
                 cost(["StringInputShouldNotWork"])
 
             # Compute after updating nominal capacity